feat(ladfile): improve globals in LAD format (#372)
# Summary
For now globals in the ladfile didn't really support values other than
"static" instances, i.e. you could only place a `LadTypeId` in the
global namespace, meaning it would have to be a pure type, and not
something like `Vec<T>`.

This change allows anything in the globals that would be allowed in
function arguments.

This also improves the formatting in the globals section:
- Types mentioned link to their detail sections
- Static instances are separated from non-static instances, to make it
easier to read
- Descriptive text is provided


![image](https://github.com/user-attachments/assets/1e9d4a95-5500-4001-8ef2-7372ae3dc56b)