feat: :sparkles: Dynamic Script Components, `register_new_component` binding, `remove_component` no longer requires `ReflectComponent` data (#379)
# Summary
In order to support dynamic script components we need:
- to relax some requirements on having a `ReflectComponent` registration
to allow component operations
- add a script component registry, which stores metadata about each
dynamically registered component
- Add `ScriptValue` to `into_script_ref` and `from_script_ref`
implementations to allow any value to be inserted into the components
payload

The nice thing about this implementation is it *just* works with the
reference system.

The current problems are that the global type cache will not reflect
this script component, to fix that we'd need a mechanism for
"re-computing" certain globals for scripts. This also does not allow
setting a schema on these, which is handy but also footgunny

As a side effect I've also made `remove_component` no longer require the
`ReflectComponent` registration