feat: allow the conversion of lua functions into `ScriptValue` via `DynamicScriptFunction` (#396)
# Summary
Because `mlua::Function`'s store a pointer to their lua context, it's
possible to call these functions without direct access to a lua context.
This allows us to convert a lua function into a `DynamicScriptFunction`
which can be marshalled as a `ScriptValue` and arbitrarilly stored on
the rust side. However this will cause issues if your script contexts
get unloaded, in the form of panics most likely