feat: improve errors when entity is unavailable (#410)
# Summary
Many places in BMS use `Entity::from_raw(0)` as an invalid entity,
because that index is not naturally assigned to any entity in Bevy.

When an entity is not available in a callback, i.e. `on_script_loaded`
people often try to interact with the entity global, this PR makes it so
that using those entities, will throw a more meaningful error and point
people to this fact.

Also fixes minor newline issue in docs